# specs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Servicing API",
    "description": "API to get service details about customers"
  },
  "servers": [
    {
      "url": "host"
    }
  ],
...
  "paths": {
...
    "/customer_accounts/{id}/service_accounts": {
      "get": {
        "tags": [
          "customer_accounts",
          "service_accounts"
        ],
        "summary": "Retrieves service accounts for a customer id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of customer account to get service accounts for",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "200 returns a service account",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/service_account"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_list"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "No Service Accounts not found"
          }
        }
      }
    },
...
    "/service_accounts/{id}": {
      "get": {
        "tags": [
          "service_accounts"
        ],
        "summary": "Retrieves service account information by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of service account to fetch",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "200 returns a service account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/service_account"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_list"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Service Account not found"
          }
        }
      }
    },
    "/service_accounts/{id}/autopay": {
      "get": {
        "tags": [
          "service_accounts",
          "payments"
        ],
        "summary": "Retrieves autopay details service account id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of service account to fetch autopay details",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "200 returns a service account",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "Id": {
                      "type": "string"
                    },
                    "StartDate": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "StopDate": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "PayMethodId": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_list"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Service Account not found"
          }
        }
      },
      "post": {
        "tags": [
          "service_accounts",
          "payments"
        ],
        "summary": "Sign up for autopay",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of service account to sign up in autopay",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/autopay_request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "200 returns a service account",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "Id": {
                      "type": "string"
                    },
                    "StartDate": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "StopDate": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "PayMethodId": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_list"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Service Account not found"
          }
        }
      },
      "delete": {
        "tags": [
          "service_accounts",
          "payments"
        ],
        "summary": "Cancels autopay for service account id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of service account to cancel autopay for",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Autopay canceled"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_list"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Service Account or autopay not found"
          }
        }
      }
    },
...
    "/service_accounts/{id}/invoices": {
      "get": {
        "tags": [
          "service_accounts",
          "invoices"
        ],
        "summary": "Retrieves invoices by service account id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of service account to fetch invoices for",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "200 returns a list of invoices",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/invoice"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_list"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Service Account or Invoices Not Found"
          }
        }
      }
    },
    "/service_accounts/{id}/invoices/{invoice_id}": {
      "get": {
        "tags": [
          "service_accounts",
          "invoices"
        ],
        "summary": "Retrieve a single invoice by service account id and invoice id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of service account to fetch invoice for",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "invoice_id",
            "in": "path",
            "description": "id of invoice to fetch",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "200 returns an invoice",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/invoice"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_list"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Service Account or Invoices Not Found"
          }
        }
      }
    },
    "/service_accounts/{id}/invoices/{invoice_id}/details": {
      "get": {
        "tags": [
          "service_accounts",
          "invoices"
        ],
        "summary": "Retrieve a single invoice by service account id and invoice id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of service account to fetch invoice for",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "invoice_id",
            "in": "path",
            "description": "id of invoice to fetch",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "200 returns a list of invoice line items",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/invoice_line_item"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_list"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Service Account or Invoice Not Found"
          }
        }
      }
    },
    "/service_accounts/{id}/invoices/{invoice_id}/meter_read_details": {
      "get": {
        "tags": [
          "service_accounts",
          "invoices"
        ],
        "summary": "Retrieves meter read details for an invoice by service account id and invoice id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of service account to fetch read details for",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "invoice_id",
            "in": "path",
            "description": "id of invoice to fetch read details",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "200 returns a list of invoice meter read details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/invoice_meter_read_detail"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_list"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Service Account or Invoice Not Found"
          }
        }
      }
    },
    "/service_accounts/{id}/invoices/{invoice_id}/download": {
      "get": {
        "tags": [
          "service_accounts",
          "invoices"
        ],
        "summary": "Downloads a single invoice by service account id and invoice id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of service account to fetch invoice for",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "invoice_id",
            "in": "path",
            "description": "id of invoice to download",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "200 returns an invoice",
            "content": {
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_list"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Service Account or Invoice Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "autopay_request": {
        "type": "object",
        "properties": {
          "pay_method_id": {
            "type": "string"
          }
        }
      },
...
      "error_list": {
        "type": "array",
        "items": {
          "type": "string"
        }
      },
      "invoice": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "string"
          },
          "Invoice_Date": {
            "type": "string",
            "format": "date"
          },
          "Due_Date": {
            "type": "string",
            "format": "date"
          },
          "Usage": {
            "type": "integer"
          },
          "Current_Charges": {
            "type": "number"
          },
          "Amount_Due": {
            "type": "number"
          },
          "Credit": {
            "type": "number"
          },
          "Balance_Forward": {
            "type": "number"
          }
        }
      },
      "invoice_line_item": {
        "type": "object",
        "properties": {
          "Description": {
            "type": "string"
          },
          "Amount": {
            "type": "number"
          },
          "UOM": {
            "type": "string"
          },
          "Quantity": {
            "type": "string"
          },
          "Multiplier": {
            "type": "number"
          },
          "Min": {
            "type": "integer"
          },
          "Max": {
            "type": "integer"
          },
          "Bill_Line_Item_Type": {
            "type": "string",
            "enum": [
              "GEXA",
              "TDSP",
              "DEPOSITS",
              "TAX"
            ]
          },
          "Bill_Line_Item_Sub_Type": {
            "type": "string",
            "enum": [
              "None",
              "Energy"
            ]
          }
        }
      },
      "invoice_meter_read_detail": {
        "type": "object",
        "properties": {
          "Invoice_Id": {
            "type": "string"
          },
          "From_Date": {
            "type": "string",
            "format": "date"
          },
          "To_Date": {
            "type": "string",
            "format": "date"
          },
          "Read_Type": {
            "type": "string"
          },
          "From_Read": {
            "type": "string"
          },
          "To_Read": {
            "type": "string"
          },
          "Usage": {
            "type": "string"
          }
        }
      },
...
      "service_account": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "string"
          },
          "Customer_Account_Id": {
            "type": "string"
          },
          "Past_Due": {
            "type": "number"
          },
          "Current_Due": {
            "type": "number"
          },
          "Deposit_Balance": {
            "type": "number"
          },
          "UAN": {
            "type": "string"
          },
          "TDU_Name": {
            "type": "string"
          },
          "TDU_DUNS_Number": {
            "type": "string"
          },
          "Service_Address": {
            "type": "object",
            "properties": {
              "Line_1": {
                "type": "string"
              },
              "Line_2": {
                "type": "string"
              },
              "City": {
                "type": "string",
                "pattern": "^[A-Za-z\\s\\-\\.]*$"
              },
              "State": {
                "type": "string",
                "enum": [
                  "AL",
                  "AK",
                  "AR",
                  "AZ",
                  "CA",
                  "CO",
                  "CT",
                  "DC",
                  "DE",
                  "FL",
                  "GA",
                  "HI",
                  "IA",
                  "ID",
                  "IL",
                  "IN",
                  "KS",
                  "KY",
                  "LA",
                  "MA",
                  "MD",
                  "ME",
                  "MI",
                  "MN",
                  "MO",
                  "MS",
                  "MT",
                  "NC",
                  "NE",
                  "NH",
                  "NJ",
                  "NM",
                  "NV",
                  "NY",
                  "ND",
                  "OH",
                  "OK",
                  "OR",
                  "PA",
                  "RI",
                  "SC",
                  "SD",
                  "TN",
                  "TX",
                  "UT",
                  "VT",
                  "VA",
                  "WA",
                  "WI",
                  "WV",
                  "WY"
                ]
              },
              "Zip": {
                "type": "string",
                "pattern": "^\\d{5}$"
              },
              "Zip_4": {
                "type": "string",
                "pattern": "^\\d{4}$"
              }
            },
            "required": [
              "Line_1",
              "City",
              "State",
              "Zip"
            ]
          },
          "Mailing_Address": {
            "type": "object",
            "properties": {
              "Line_1": {
                "type": "string"
              },
              "Line_2": {
                "type": "string"
              },
              "City": {
                "type": "string",
                "pattern": "^[A-Za-z\\s\\-\\.]*$"
              },
              "State": {
                "type": "string",
                "enum": [
                  "AL",
                  "AK",
                  "AR",
                  "AZ",
                  "CA",
                  "CO",
                  "CT",
                  "DC",
                  "DE",
                  "FL",
                  "GA",
                  "HI",
                  "IA",
                  "ID",
                  "IL",
                  "IN",
                  "KS",
                  "KY",
                  "LA",
                  "MA",
                  "MD",
                  "ME",
                  "MI",
                  "MN",
                  "MO",
                  "MS",
                  "MT",
                  "NC",
                  "NE",
                  "NH",
                  "NJ",
                  "NM",
                  "NV",
                  "NY",
                  "ND",
                  "OH",
                  "OK",
                  "OR",
                  "PA",
                  "RI",
                  "SC",
                  "SD",
                  "TN",
                  "TX",
                  "UT",
                  "VT",
                  "VA",
                  "WA",
                  "WI",
                  "WV",
                  "WY"
                ]
              },
              "Zip": {
                "type": "string",
                "pattern": "^\\d{5}$"
              },
              "Zip_4": {
                "type": "string",
                "pattern": "^\\d{4}$"
              }
            },
            "required": [
              "Line_1",
              "City",
              "State",
              "Zip"
            ]
          },
          "Paperless_CEN": {
            "type": "boolean"
          },
          "Paperless_Billing": {
            "type": "boolean"
          },
          "Budget_Billing": {
            "type": "boolean"
          },
          "Service_Stop_Request_date": {
            "type": "string",
            "format": "date-time"
          },
          "Status": {
            "type": "string",
            "enum": [
              "None",
              "Signed_Up",
              "Active",
              "Temporarily_Inactive",
              "Disconnected",
              "Suspended",
              "Voided",
              "Unapproved",
              "Rejected",
              "Approved",
              "Closed"
            ]
          },
          "Terminate_Switch": {
            "type": "boolean"
          },
          "Switch_Hold": {
            "type": "boolean"
          },
          "Service_Start_Date": {
            "type": "string",
            "format": "date-time"
          },
          "Service_Stop_Date": {
            "type": "string",
            "format": "date-time"
          },
          "Contract_Start_Date": {
            "type": "string",
            "format": "date-time"
          },
          "Contract_End_Date": {
            "type": "string",
            "format": "date-time"
          },
          "Is_Auto_Bill_Pay": {
            "type": "boolean"
          },
          "PayMethodId": {
            "type": "integer"
          },
          "AutoPayConfigId": {
            "type": "integer"
          },
          "Is_Happiness_Guaranteed": {
            "type": "boolean"
          },
          "Average_Usage": {
            "type": "integer"
          },
          "Zone": {
            "type": "string"
          },
          "Usage_Profile": {
            "type": "string",
            "enum": [
              "Low",
              "Medium",
              "High",
              "Ultra_High"
            ]
          },
          "Current_Offer": {
            "type": "object",
            "properties": {
              "Rate_Code": {
                "type": "string"
              },
              "Client_Key": {
                "type": "string"
              },
              "Product_Name": {
                "type": "string"
              },
              "Product_Description": {
                "type": "string"
              },
              "RateAt500kwh": {
                "type": "number"
              },
              "RateAt1000kwh": {
                "type": "number"
              },
              "RateAt2000kwh": {
                "type": "number"
              },
              "Term": {
                "type": "integer"
              },
              "TDU_Name": {
                "type": "string"
              },
              "TDU_DUNS_Number": {
                "type": "string"
              },
              "Meter_Charge": {
                "type": "number"
              },
              "Delivery_Charge": {
                "type": "number"
              },
              "Family": {
                "type": "string",
                "enum": [
                  "LEVEL_1",
                  "LEVEL_2",
                  "LEVEL_3",
                  "LEVEL_4",
                  "LEVEL_5"
                ]
              },
              "Type": {
                "type": "string",
                "enum": [
                  "Acquisition",
                  "Renewal"
                ]
              },
              "Early_Termination_Fee": {
                "type": "number"
              },
              "Base_Charge": {
                "type": "number"
              },
              "Energy_Charges": {
                "type": "array",
                "minItems": 1,
                "items": {
                  "type": "object",
                  "properties": {
                    "Min": {
                      "type": "integer"
                    },
                    "Max": {
                      "type": "integer"
                    },
                    "Amount": {
                      "type": "number"
                    }
                  }
                }
              },
              "Usage_Credits": {
                "type": "array",
                "minItems": 1,
                "items": {
                  "type": "object",
                  "properties": {
                    "Min": {
                      "type": "integer"
                    },
                    "Max": {
                      "type": "integer"
                    },
                    "Amount": {
                      "type": "number"
                    }
                  }
                }
              },
              "Usage_Charges": {
                "type": "array",
                "minItems": 1,
                "items": {
                  "type": "object",
                  "properties": {
                    "Min": {
                      "type": "integer"
                    },
                    "Max": {
                      "type": "integer"
                    },
                    "Amount": {
                      "type": "number"
                    }
                  }
                }
              },
              "Start_Date": {
                "type": "string",
                "format": "date-time"
              },
              "End_Date": {
                "type": "string",
                "format": "date-time"
              },
              "Partner_Info": {
                "type": "array",
                "minItems": 1,
                "items": {
                  "type": "object",
                  "properties": {
                    "Partner": {
                      "type": "object",
                      "properties": {
                        "Id": {
                          "type": "string"
                        },
                        "Name": {
                          "type": "string"
                        },
                        "Description": {
                          "type": "string"
                        },
                        "Billing_Code": {
                          "type": "string"
                        },
                        "Creation_Time": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "Date_Created": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "Date_Last_Modified": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    },
                    "Code": {
                      "type": "string"
                    }
                  }
                }
              },
              "Featured_Usage_Level": {
                "type": "string"
              },
              "IsHoldOverRate": {
                "type": "boolean"
              },
              "IsLegacyOffer": {
                "type": "boolean"
              },
              "IsFixed": {
                "type": "boolean"
              },
              "Has_Partner": {
                "type": "boolean"
              },
              "Cost_Components": {
                "type": "array",
                "minItems": 1,
                "items": {
                  "type": "object",
                  "properties": {
                    "Min": {
                      "type": "integer"
                    },
                    "Max": {
                      "type": "integer"
                    },
                    "Amount": {
                      "type": "number"
                    },
                    "Operation_Type": {
                      "type": "string",
                      "enum": [
                        "Additive",
                        "Multiplicative"
                      ]
                    },
                    "IsCompound": {
                      "type": "boolean"
                    },
                    "TimeOfUseInterval": {
                      "type": "string"
                    },
                    "DisplayContent": {
                      "type": "string"
                    }
                  }
                }
              },
              "Is_BOGO": {
                "type": "boolean"
              },
              "Is_TOU": {
                "type": "boolean"
              },
              "Is_RTP": {
                "type": "boolean"
              }
            }
          },
          "Offering_Name": {
            "type": "string"
          },
          "Tos_Fee_Id": {
            "type": "integer"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}